Fix `Settings::get_baud_rate` on Linux & Android
The `cfgetospeed` function is not supposed to be called
with a Termios2 struct. This leads to undefined behavior,
depending on the libc implementation. The issue probably
got triggered by a recent commit to glibc (5cf101a).

As the Termios2 API used on Linux & Android always sets
the `c_ospeed` field to the exact baud rate, no matter
if the baud rate is one of the predefined values, just use
the `c_ospeed` value directly.
The predefined baud rates can therefore also be removed.